Updates int32 to int (#55)
* Map `pg_catalog.int4` to `int` in SQLC configuration

* Change `GetIntFromRequestPath` to return `int` instead of `int32` and simplify path variable parsing logic.

* Update `Querier` interface and related code to use `int` instead of `int32` for IDs, align tests, request structs, and SQLC-generated files.

* Switch `int32` to `int` for IDs across product-related codebase, including SQLC-generated files, interfaces, tests, request structs, and related methods.

+++ internal/product/request_test.go
@@ -58,7 +58,7 @@ func TestUpdateProductRequest_ToParams(t *testing.T) {
 		Name:        "Updated Product",
 		Description: "Updated Description",
 	}
-	id := int32(10)
+	id := 10
 
 	params := req.ToParams(id)
